+++ alicloud/resource_alicloud_route_table.go
@@ -2,10 +2,11 @@ package alicloud
 
 import (
 	"fmt"
+	"strings"
+
 	"github.com/aliyun/alibaba-cloud-sdk-go/services/vpc"
 	"github.com/hashicorp/terraform/helper/schema"
 	"github.com/terraform-providers/terraform-provider-alicloud/alicloud/connectivity"
-	"strings"
 )
 
 func resourceAliyunRouteTable() *schema.Resource {